feat: DACVAE????FT?FP32??????????????fp32?
- train.cli.finetune_dacvae_decoder ???????????float32???
- train_500m_v3_full / TrainConfig / CLI ? precision ? fp32 ???

+++ core/config.py
@@ -175,7 +175,7 @@ class TrainConfig:
     valid_split_by: str = "sample"
     progress: bool = True
     progress_all_ranks: bool = False
-    precision: str = "bf16"
+    precision: str = "fp32"
     grad_clip_norm: float = 1.0
     gradient_accumulation_steps: int = 1
     max_text_len: int = 256
